requirements.txt
ipykernel
langchain
langchain_core
langchain_community
langchain_groq
langgraph
mysql-connector-python
geocoder
python-dotenv
mcp-use
bcyrpt